Batching rule for torch.mul
This PR implements the batching rule for `torch.mul` for the (Tensor,
Tensor) overload.

NB: ~250 lines of this PR are tests, so please don't be scared away by
the line count.

It introduces the BroadcastingVmapTransform, which is the VmapTransform
one should use for operations that broadcast their inputs. This
transform:
- permutes all batch dimensions to the front of the tensors
- aligns the batch dimensions of the tensors, adding extra 1's where
necessary
- aligns the non-batch dims of the tensors, adding extra 1's where
necessary.

Test Plan:
- Test BroadcastingVmapTransform in `./build/bin/vmap_test`.
- Test mul_batching_rule in `./build/bin/vmap_test`.

ghstack-source-id: 35b42de45e38cb318826f2d10f0d6834fe1d8608
Pull Request resolved: https://github.com/pytorch/pytorch/pull/39859