Implement tensor.size(int) for BatchedTensor
We have this call native::size directly. Some alternatives I considered
were:
- Call VariableType::size directly. That seems isomorphic to what we're
doing now.
- when creating a BatchedTensor from a regular tensor, put all of the
keys on that tensor into the BatchedTensor's dispatch key set and use
the dispatcher fallthrough mechanism. That seems weird because
BatchedTensor is a tensor wrapper and also error prone because if
BatchedTensor gets the VariableType key, there's a chance that if
something goes wrong, an autogradmeta gets created on it...

Test Plan:
- `./build/bin/vmap_test`

ghstack-source-id: 564d0889da079e32a52fda2e214bea6387222044
Pull Request resolved: https://github.com/pytorch/pytorch/pull/40028