Add batching rule for torch.expand
This is (probably) necessary for the vmap frontend API (coming up after
this PR should be the vmap frontend API).

There is some manual handling of sizes in the `expand_batching_rule`.
In particular, when performing expand(Tensor[B0, 3], [2, 3]), where B0
is a batch dimension and Tensor[B0, 3] is a batched tensor with batch
dimension B0, we can't call expand directly on the physical view and
instead first need to perform a view.
It's possible to add said view as a helper function on `VmapPhysicalView` but
after reading through the operator spreadsheet the conclusion was that
no other operator needs the same manual handling.

Test Plan:
- `./build/bin/vmap_test`

ghstack-source-id: 27d24ec8d2aa51cdd2bed0641df299ba005481d8
Pull Request resolved: https://github.com/pytorch/pytorch/pull/40097